LayoutTests/inspector/debugger/break-on-exception-throw-in-promise-rethrow-in-catch-expected.txt

CONSOLE MESSAGE: line 80: testThrowingInPromiseWithRethrowInCatch
CONSOLE MESSAGE: line 82: in promise
CONSOLE MESSAGE: line 7: throwing TestError
CONSOLE MESSAGE: line 85: in promise.catch
Checking pause locations when pausing on all exceptions.

PAUSE #1 AT: doThrow:8:22
PAUSE #2 AT: promiseCatch:86:16
